third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/elements-panel-search-expected.txt

Tests that elements panel search is returning proper results.


Running: testSetUp

Running: testPlainText
F o o B a r

Running: testPartialText
F o o B a r

Running: testStartTag
< i n p u t   v a l u e = " I n p u t V a l " >

Running: testEndTag
< i n p u t   v a l u e = " I n p u t V a l " >

Running: testPartialTag
< i n p u t   v a l u e = " I n p u t V a l " >

Running: testPartialAbsentTagStart
Nothing found

Running: testPartialAbsentTagEnd
Nothing found

Running: testFullTag
< i n p u t   v a l u e = " I n p u t V a l " >

Running: testExactAttributeName
< i n p u t   v a l u e = " I n p u t V a l " >

Running: testExactAttributeVal_ue
< i n p u t   v a l u e = " I n p u t V a l " >

Running: testExactAttributeVal_ueOnRoot
< h t m l   i d = " d o c u m e n t E l e m e n t " >

Running: testExactAttributeVal_ueWithQuotes
< d i v   a t t r = " f o o " > < / d i v >

Running: testPartialAttributeVal_ue
< i n p u t   v a l u e = " I n p u t V a l " >

Running: testXPathAttribute
< d i v   a t t r = " f o o " > < / d i v >

Running: testSelector
< s p a n > F o u n d   b y   s e l e c t o r < / s p a n >

Running: testCaseUpperFindsLower
< s p a n   c l a s s = " f o o   k o o " > < / s p a n >

Running: testCaseLowerFindsUpper
< s p a n   c l a s s = " C A S E L E S S " > < / s p a n >

Running: testCaseAttribute
< s p a n   d a t a - c a m e l = " i n s e n s t i v e " > < / s p a n >

Running: testSearchShadowDOM
< c o n t e n t > < / c o n t e n t >

Running: testSearchUAShadowDOM
Searching UA shadow DOM with setting disabled:
Nothing found
Searching UA shadow DOM with setting enabled:
< d i v   i d = " u a - s h a d o w - c o n t e n t " > < / d i v >

Running: testSearchShadowHostChildren
< d i v   i d = " s h a d o w - h o s t - c o n t e n t " > < / d i v >

Running: testSearchClosingTag
< t e x t a r e a > < / t e x t a r e a >
